+++ DistributedDatabaseService.DAL/LabDbContext.cs
@@ -11,9 +11,13 @@ namespace DistributedDatabaseService.DAL
         public DbSet<Student> Students { get; set; }
         public DbSet<Course> Courses { get; set; }
         public DbSet<StudentCourse> StudentCourses { get; set; }
+        public DbSet<Group> Groups { get; set; }
+        public DbSet<Faculty> Faculties { get; set; }
+        public DbSet<Instructor> Instructors { get; set; }
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
+            // StudentCourse - багато до багатьох
             modelBuilder.Entity<StudentCourse>()
                 .HasKey(sc => new { sc.StudentId, sc.CourseId }); // Композитний ключ
 
@@ -26,7 +30,30 @@ namespace DistributedDatabaseService.DAL
                 .HasOne(sc => sc.Course)
                 .WithMany(c => c.StudentCourses)
                 .HasForeignKey(sc => sc.CourseId);
+
+            // Student - Group (багато до одного)
+            modelBuilder.Entity<Student>()
+                .HasOne(s => s.Group)
+                .WithMany(g => g.Students)
+                .HasForeignKey(s => s.GroupId);
+
+            // Group - Faculty (багато до одного)
+            modelBuilder.Entity<Group>()
+                .HasOne(g => g.Faculty)
+                .WithMany(f => f.Groups)
+                .HasForeignKey(g => g.FacultyId);
+
+            // Instructor - Faculty (багато до одного)
+            modelBuilder.Entity<Instructor>()
+                .HasOne(i => i.Faculty)
+                .WithMany(f => f.Instructors)
+                .HasForeignKey(i => i.FacultyId);
+
+            // Course - Instructor (багато до одного)
+            modelBuilder.Entity<Course>()
+                .HasOne(c => c.Instructor)
+                .WithMany(i => i.Courses)
+                .HasForeignKey(c => c.InstructorId);
         }
     }
 }
-